added replay feature
added a command line option `--replay` that will replay the state "live"
as it happened.  Updates will come in as they did during the actual live
feed delayed by the appropriate time between events.  If the feed
catches up to real-time then it will continue polling and displaying
updates as usual.  During the replay, if the stage is still in progress,
polling will continue and new events will be added to the queue.