Added a status command and various bugfixes
Added a status command (`s`) to show you some info about the entries in
the backlog.  While doing that, refactored the keypress handling a bit,
which led to some bugs, and to some bugfixes.

Most notably was the fact that when a module function is called as an
event handler:

e.g. `process.stdin.on('keypress', module.someFunc)`

or to execute on the next event loop tick:

e.g. `process.nextTick(module.someFunc)`

The module context doesn't remain, so in `module.someFunc`, the value of
`this` doesn't refer to the module anymore.

It had been working fine when the `next()` method and event handlers
were all defined in the `tdfLive.start()` method, and they were all
referencing `self` (which stored the value of `this` from the original
invocation).

once the `next()` method was moved out into the module, as well as the
keypress handling, then some errors crept in because those methods lost
their context.  The fix was to `bind()` the module context to the
callback function:

e.g. `process.nextTick(module.someFunc.bind(module))`

Also added some notes into the README about the new `--replay` feature
and keystrokes.