fix showing the latest timestamp in the status
if the feed is polling, and there aren't any entries in the backlog,
then there was a bug when you displayed the status.  it would crash.
bad bug.  because it was trying to get the latest timestamp by looking
at the last entry, but if there are no entries then it'd cause an error.

fix was to save the status in a property of the object, updated whenever
a new updates was received.  Then use that to display the timestamp in
the status output.